% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_dist.R
\name{conv_dist}
\alias{conv_dist}
\title{Convert metric units of distance between each other. The basic unit of distance is centimeters, expressed as 'cm'. Examples listed below.}
\usage{
conv_dist(num, from, to)
}
\arguments{
\item{num}{The decimal number units of the 'from' units. Can also pass as vector.}

\item{from}{The initial units before conversion, taken as a character. Includes: ft, m, mi, yard, in}

\item{to}{The desired units after conversion, taken as a character. Includes: ft, m, mi, yard, in}
}
\value{
A decimal (or vector) displaying the number of 'to' units
}
\description{
Convert metric units of distance between each other. The basic unit of distance is centimeters, expressed as 'cm'. Examples listed below.
}
\examples{
conv_dist(15, 'ft', 'm')
conv_dist(60, 'yd', 'm' )
}
