% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_mass.R
\name{conv_mass}
\alias{conv_mass}
\title{Convert between kilograms, pounds, ounces, and stones.}
\usage{
conv_mass(num, from = "", to = "")
}
\arguments{
\item{num}{the amount of the unit you want to onvert.}

\item{from}{the original unit.}

\item{to}{the unit you want to convert to}
}
\value{
data frame of converted value and unit.
}
\description{
Convert between kilograms, pounds, ounces, and stones.
}
\details{
num: a positive number

from: the original unit. To enter kilograms type "kg" or "kilogram". To enter pounds type "lb" or "pound". To enter ounces type "oz" or "ounce". To enter stones type "st" or "stone"

to: the unit you want to convert to. To enter kilograms type "kg" or "kilogram". To enter pounds type "lb" or "pound". To enter ounces type "oz" or "ounce". To enter stones type "st" or "stone"
}
\examples{
conv_mass(4, "kg", "lbs")
conv_mass(25, "st", "ounces")
}
