% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_temp.R
\name{conv_temp}
\alias{conv_temp}
\title{Convert between between celsius, kelvin and fahrenheit.}
\usage{
conv_temp(num, from = "", to = "")
}
\arguments{
\item{num}{the amount of the unit you want to onvert.}

\item{from}{the original unit.}

\item{to}{the unit you want to convert to}
}
\value{
data frame of converted value and unit.
}
\description{
Convert between between celsius, kelvin and fahrenheit.
}
\details{
num: a positive number

from: the original unit. To enter celsius type "c" or "celsius". To enter kelvin type "k" or "kelvin". To enter fahrenheit type "f" or "fahrenheit".

to: the unit you want to convert to. To enter celsius type "c" or "celsius". To enter kelvin type "k" or "kelvin". To enter fahrenheit type "f" or "fahrenheit".
}
\examples{
conv_temp(30, "c", "kelvin")
conv_tem(100, "f", "celsius")
}
