% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conv_vol.R
\name{conv_volume}
\alias{conv_volume}
\title{Convert between liters, gallons, quarts, and pints.}
\usage{
conv_volume(num, from = "", to = "")
}
\arguments{
\item{num}{the amount of the unit you want to onvert.}

\item{from}{the original unit.}

\item{to}{the unit you want to convert to}
}
\value{
data frame of converted value and unit.
}
\description{
Convert between liters, gallons, quarts, and pints.
}
\details{
num: a positive number

from: the original unit. To enter liters type "l" or "liter". To enter gallons type "gal" or "gallon". To enter quarts type "qt" or "quart". To enter pints type "pt" or "pint"

to: the unit you want to convert to. To enter liters type "l" or "liter". To enter gallons type "gal" or "gallon". To enter quarts type "qt" or "quart". To enter pints type "pt" or "pint"
}
\examples{
conv_volume(4, "l", "gal")
conv_volume(25, "pint", "qt")
}
